fix: address review feedback on process output panel (#252)
Security and type-safety pass over the panel introduced in #1172:

- Skip routing PASS_GREP through the panel. pass grep returns lines
  from password files (the very point of the command), so streaming
  them into a long-lived UI panel is a privacy leak. Also skip
  PASS_INSERT defensively: stdout normally carries gpg/git progress
  only, but a future code path with --echo or similar would leak the
  stdin password.
- Type the finishedAnyWithPid signal as Enums::PROCESS instead of
  int. Drop the static_cast at the emit site and switch the receiver
  lambda to take Enums::PROCESS so comparisons are type-checked
  against the enum domain. Add Q_DECLARE_METATYPE(Enums::PROCESS) so
  queued connections (any future cross-thread use) work.
- Make MainWindow::getProcessName a static helper (no instance state
  is touched) and let it accept Enums::PROCESS directly. Drop the
  unused int pid parameter from appendProcessOutput; only
  onProcessOutput needs the pid to pick a label prefix.
- Wrap the command labels (`git init`, `pass insert`, …) in
  QStringLiteral with explicit `// no-tr` comments. They reference
  the actual shell commands a user would type to reproduce the
  action, so they should not be localized; the comment marks the
  choice as intentional. Cover GIT_MOVE / GIT_COPY too (previously
  fell into the default branch and produced an empty prefix).
- Replace the `default:` arm of the switch with explicit non-emit
  cases for PASS_SHOW, PASS_OTP_GENERATE, PROCESS_COUNT, INVALID
  so the compiler will warn if a new Enums::PROCESS member is added
  without a label.
- Add the missing trailing newline at end of mainwindow.cpp.

+++ src/pass.cpp
@@ -713,7 +713,7 @@ auto Pass::formatInsertError(const QString &friendly, const QString &err)
 void Pass::emitProcessFinishedSignal(PROCESS pid, const QString &out,
                                      const QString &err) {
   emit finishedAny(out, err);
-  emit finishedAnyWithPid(out, err, static_cast<int>(pid));
+  emit finishedAnyWithPid(out, err, pid);
 
   switch (pid) {
   case GIT_INIT:
